修改工具合集名称及优化Navigate切换代码
