+++ SYSTools/Pages/Toolkit.xaml
@@ -1,4 +1,4 @@
-﻿<Page x:Class="SYSTools.Pages.Other"
+﻿<Page x:Class="SYSTools.Pages.Toolkit"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -8,7 +8,7 @@
       mc:Ignorable="d" 
       d:DesignHeight="610" d:DesignWidth="1040"
       KeepAlive="True"
-      Title="Other" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" MinWidth="1040" MinHeight="610" Loaded="Window_Loading">
+      Title="Toolkit" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" MinWidth="1040" MinHeight="610" Loaded="Window_Loading">
 
     <Grid>
